ocm: Split resources into files
Each file represents operations on specific OCM resources while being
part of the same package. This simplifies code structure and provides
a logical place for finding and adding new resource operations.
