+++ pkg/ocm/idps.go
@@ -1,0 +1,53 @@
+/*
+Copyright (c) 2021 Red Hat, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+  http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package ocm
+
+import (
+	cmv1 "github.com/openshift-online/ocm-sdk-go/clustersmgmt/v1"
+)
+
+func GetIdentityProviders(client *cmv1.ClustersClient, clusterID string) ([]*cmv1.IdentityProvider, error) {
+	idpClient := client.Cluster(clusterID).IdentityProviders()
+	response, err := idpClient.List().
+		Page(1).
+		Size(-1).
+		Send()
+	if err != nil {
+		return nil, handleErr(response.Error(), err)
+	}
+
+	return response.Items().Slice(), nil
+}
+
+func IdentityProviderType(idp *cmv1.IdentityProvider) string {
+	switch idp.Type() {
+	case "GithubIdentityProvider":
+		return "GitHub"
+	case "GitlabIdentityProvider":
+		return "GitLab"
+	case "GoogleIdentityProvider":
+		return "Google"
+	case "HTPasswdIdentityProvider":
+		return "htpasswd"
+	case "LDAPIdentityProvider":
+		return "LDAP"
+	case "OpenIDIdentityProvider":
+		return "OpenID"
+	}
+
+	return ""
+}
